requirements.txt
numpy
scipy
matplotlib
pillow
streamlit
streamlit-drawable-canvas
